merge:get code
